Add polling trait
